+++ R/spatial_enrichment_visuals.R
@@ -1,0 +1,146 @@
+
+#' @title plotCellTypesFromEnrichment
+#' @name plotCellTypesFromEnrichment
+#' @param gobject Giotto Object
+#' @param spat_unit spatial unit in which the enrichment information is stored
+#' @param feat_type feature type for which the enrichment information was calculated
+#' @param enrichment_name name of the spatial enrichment
+#'  i.e. output from GiottoClass::list_spatial_enrichment_names()
+#'  Default value is "PAGE_Z_score"
+#' @param return_frequency_table see details. Default FALSE
+#' @details 
+#' This function returns a two-column matrix, one column
+#' will contain cell IDs and the other will contain the 
+#' most likely cell type based on the provided enrichment.
+#' 
+#' By setting return_frequency_table to TRUE, a table
+#' will instead be returned. This table will detail
+#' the number of occurences for each cell type, based on
+#' the provided enrichment.
+#' 
+#' The cell types are assigned by determining the maximum 
+#' value of the z-score or -log10(p-value) for a cell and
+#' the associated cell types from the enrichment.
+#' 
+#' @export 
+findCellTypesFromEnrichment <- function(gobject = NULL,
+                                        spat_unit = NULL,
+                                        feat_type = NULL,
+                                        enrichment_name = "PAGE_z_score",
+                                        return_frequency_table = FALSE){
+    # guard clauses
+
+    if(!inherits(gobject, "giotto")) stop("gobject needs to be a giotto object")
+
+    spat_unit = set_default_spat_unit(gobject = gobject,
+                                      spat_unit = spat_unit)
+    feat_type = set_default_feat_type(gobject = gobject,
+                                      spat_unit = spat_unit,
+                                      feat_type = feat_type)
+
+    # data.table variables
+    probable_cell_type = cell_ID = NULL
+
+    # extract p-value or z-socre from provided enrichment
+    pz_enrich = getSpatialEnrichment(gobject,
+                                     spat_unit = spat_unit,
+                                     feat_type = feat_type, 
+                                     name = enrichment_name,
+                                     output = "data.table")
+
+    if( colnames(pz_enrich)[[1]] != "cell_ID"){
+        selected_cols = colnames(pz_enrich)[colnames(pz_enrich) != "cell_ID"]
+        setcolorder(pz_enrich, c("cell_ID", selected_cols))
+    }
+
+
+    n_c = ncol(pz_enrich)
+
+    # Find the cell type column that corresponds to the
+    # maximum value within a row and assign it into a
+    # new column, mapping a cell to it's most likely type
+    pz_enrich[, probable_cell_type :=  names(.SD)[max.col(.SD)], .SDcols = 2:n_c]
+
+    cell_ID_and_types_pz_enrich = pz_enrich[, .(cell_ID, probable_cell_type)]
+
+    if(return_frequency_table) {
+        pz_enrich_cell_type_frequencies = table(cell_ID_and_types_pz_enrich$probable_cell_type)
+        return(pz_enrich_cell_type_frequencies)
+    }
+
+    return(cell_ID_and_types_pz_enrich)
+}
+
+#' @title plotCellTypesFromEnrichment
+#' @name plotCellTypesFromEnrichment
+#' @param gobject Giotto Object
+#' @param spat_unit spatial unit in which the enrichment information is stored
+#' @param feat_type feature type for which the enrichment information was calculated
+#' @param enrichment_name name of the spatial enrichment
+#'  i.e. output from GiottoClass::list_spatial_enrichment_names()
+#'  Default value is "PAGE_Z_score"
+#' @param title Title of the generated plot. Default `paste0(spat_unit,"cell types (maximum", enrichment_name, ")")`
+#' @param save_param arguments passed to all_plots_save_function; must be a list
+#' @param default_save_name name under which the plot will be saved, default  `cell_types_from_enrichment`
+#' @param save_plot binary option to save plot, default NA - will read instructions of gobject
+#' @param show_plot binary option to show plot in viewer, default NA - will read instructions of gobject
+#' @param return_plot binary option to return the plot object, default NA - will read instructions of gobject
+#' @details 
+#' 
+#' This function generates a bar plot of cell types vs the frequency
+#' of that cell type in the data. These cell type resutls are
+#' based on the provided `enrichment_name`, and will be determined
+#' by the maximum value of the z-score or p-value for a given cell or annotation.
+#' 
+#' @export
+plotCellTypesFromEnrichment <- function(gobject = NULL,
+                                        spat_unit = NULL,
+                                        feat_type = NULL,
+                                        enrichment_name = "PAGE_z_score",
+                                        title = NULL,
+                                        save_param =  list(),
+                                        default_save_name = 'cell_types_from_enrichment',
+                                        save_plot = NA,
+                                        show_plot = NA,
+                                        return_plot = NA){
+    # guard clauses handled at first step downstream
+    # therefore, omitting here.
+    id_and_types = findCellTypesFromEnrichment(gobject = gobject,
+                                               spat_unit = spat_unit,
+                                               feat_type = feat_type,
+                                               enrichment_name = enrichment_name,
+                                               return_frequency_table = FALSE)
+
+    # data.table column
+    probable_cell_type = NULL
+
+    if(is.null(title)) title = paste0(spat_unit,"cell types (maximum", enrichment_name, ")")
+    
+    pl <- ggplot2::ggplot(id_and_types, aes(x = probable_cell_type)) + 
+            ggplot2::geom_bar() +
+            ggplot2::theme(axis.text.x = element_text(angle = 45),
+                           axis.ticks.length.x =unit(1.5, "cm")) +
+            ggplot2::labs(title = title,
+                          x = "Cell Type",
+                          y = "Frequency")
+
+    # print, return and save parameters
+    show_plot = ifelse(is.na(show_plot), readGiottoInstructions(gobject, param = 'show_plot'), show_plot)
+    save_plot = ifelse(is.na(save_plot), readGiottoInstructions(gobject, param = 'save_plot'), save_plot)
+    return_plot = ifelse(is.na(return_plot), readGiottoInstructions(gobject, param = 'return_plot'), return_plot)
+    
+    ## print plot
+    if(show_plot == TRUE) {
+      print(pl)
+    }
+    ## save plot
+    if(save_plot == TRUE) {
+      do.call('all_plots_save_function', c(list(gobject = gobject, plot_object = pl, default_save_name = default_save_name), save_param))
+    }
+    ## return plot
+    if(return_plot == TRUE) {
+      return(pl)
+    }
+}
+
+