+++ NEWS.md
@@ -5,6 +5,9 @@
 ## Breaking Changes
 
 ## Added
+- New File `spatial_enrichment_visuals.R`
+- New function `findCellTypesFromEnrichment()` within `spatial_enrichment_visuals.R` to show most probable cell types based on a provided enrichment
+- New function `plotCellTypesFromEnrichment()` within `spatial_enrichment_visuals.R` that generates a bar plot of cell types vs frequency based on a provided enrichment
 
 ## Changes